# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleCreate": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "allow",
              "deny"
            ],
            "title": "Decision",
            "type": "string"
          },
          "dependency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Dependency Key",
            "type": "string"
          },
          "dependency_type": {
            "enum": [
              "agent",
              "model",
              "skill",
              "platform",
              "client"
            ],
            "title": "Dependency Type",
            "type": "string"
          },
          "dependency_version_range": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Dependency Version Range",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "reason_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "agent",
              "model",
              "skill",
              "client"
            ],
            "title": "Subject Type",
            "type": "string"
          },
          "subject_version_range": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject Version Range",
            "type": "string"
          }
        },
        "required": [
          "subject_type",
          "subject_id",
          "subject_version_range",
          "dependency_type",
          "dependency_key",
          "dependency_version_range",
          "decision",
          "reason_code",
          "priority"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
      "RuleResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "allow",
              "deny"
            ],
            "title": "Decision",
            "type": "string"
          },
          "dependency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Dependency Key",
            "type": "string"
          },
          "dependency_type": {
            "enum": [
              "agent",
              "model",
              "skill",
              "platform",
              "client"
            ],
            "title": "Dependency Type",
            "type": "string"
          },
          "dependency_version_range": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Dependency Version Range",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "reason_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "validated",
              "published"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "agent",
              "model",
              "skill",
              "client"
            ],
            "title": "Subject Type",
            "type": "string"
          },
          "subject_version_range": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject Version Range",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "subject_type",
          "subject_id",
          "subject_version_range",
          "dependency_type",
          "dependency_key",
          "dependency_version_range",
          "decision",
          "reason_code",
          "priority",
          "id",
          "status",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "RuleResponse",
        "type": "object"
      },
      "RuleUpdate": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "anyOf": [
              {
                "enum": [
                  "allow",
                  "deny"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "dependency_version_range": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Version Range"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reason_code": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "subject_version_range": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Version Range"
          }
        },
        "title": "RuleUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "iBreeze Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/api/v1/compatibility-rules": {
      "get": {
        "operationId": "list_rules_endpoint_admin_api_v1_compatibility_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Rules Endpoint Admin Api V1 Compatibility Rules Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Rules Endpoint",
        "tags": [
          "compatibility"
        ]
      },
      "post": {
        "operationId": "create_rule_endpoint_admin_api_v1_compatibility_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Rule Endpoint",
        "tags": [
          "compatibility"
        ]
      }
    },
    "/admin/api/v1/compatibility-rules/{rule_id}": {
      "delete": {
        "operationId": "delete_rule_endpoint_admin_api_v1_compatibility_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Rule Endpoint",
        "tags": [
          "compatibility"
        ]
      },
      "get": {
        "operationId": "get_rule_endpoint_admin_api_v1_compatibility_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Rule Endpoint",
        "tags": [
          "compatibility"
        ]
      },
      "patch": {
        "operationId": "update_rule_endpoint_admin_api_v1_compatibility_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Rule Endpoint",
        "tags": [
          "compatibility"
        ]
      }
    },
    "/admin/api/v1/compatibility-rules/{rule_id}/validate": {
      "post": {
        "operationId": "validate_rule_endpoint_admin_api_v1_compatibility_rules__rule_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Rule Endpoint",
        "tags": [
          "compatibility"
        ]
      }
    },
...
  }
}